chore: reorganize http files.
