+++ otel/http/server/handler_test.go
@@ -1,4 +1,4 @@
-package http
+package server
 
 import (
 	"context"
